requirements.txt
beautifulsoup4==4.12.2
bokeh==3.1.1
google-cloud-enterpriseknowledgegraph==0.3.2
numpy==1.24.3
pandas==2.0.1
requests==2.31.0
Scrapy==2.9.0
thefuzz==0.19.0
Unidecode==1.3.6